Finita pagina di Aggiunta nuovo prodotto. Cambiato il modo con cui vengono recuperate le immagini dei prodotti (tramite servlet adesso). Per funzionare bisogna settare nella configurazione di tomcat in Intellij una variabile di environment chiamata STUDENT_SUPPS con valore la path assoluta della cartella del progetto. Questo viene fatto in quanto questa potrebbe trovarsi in path diverse per ogni pc. Alla fine dello sviluppo questo verrà rimosso e rimpiazzato con una path assoluta unica.
-Vari fix, aggiunte e/o modificate alcune funzioni dei dao, aumentato il varchar del nome prodotto a 50 nel DB.
-Altre migliore sparse.

+++ src/main/java/com/tsw/studentsupps/Controller/Admin/AdminAddProductServlet.java
@@ -51,6 +51,13 @@ public class AdminAddProductServlet extends HttpServlet {
             return;
         }
 
+        if(ProdottoDAO.doExistsByName(request.getParameter("name"))) {
+            request.setAttribute("addProductStatus", "nameTaken");
+            RequestDispatcher dispatcher= request.getRequestDispatcher("/pages/Admin/AddProduct.jsp");
+            dispatcher.forward(request, response);
+            return;
+        }
+
         Prodotto p= new Prodotto();
         String name= request.getParameter("name");
         p.setNome(name);
@@ -59,14 +66,14 @@ public class AdminAddProductServlet extends HttpServlet {
         p.setIVA(Short.parseShort(request.getParameter("iva")));
         p.setQuantita(Integer.parseInt(request.getParameter("quantity")));
 
-        /* FUNZIONANTE IN PARTE*/
         Part imagePart= request.getPart("image");
-        File uploads= new File(getServletContext().getInitParameter("uploadImageProduct.location"));
+        File uploads= new File((String) getServletContext().getAttribute("prodImageFolder"));
         String imageName= imagePart.getSubmittedFileName();
         File imageFile= new File(uploads, name + imageName.substring(imageName.lastIndexOf(".")));
 
         try (InputStream input= imagePart.getInputStream()) {
-            Files.copy(input, imageFile.toPath());
+            if(!Files.exists(imageFile.toPath()))
+                Files.copy(input, imageFile.toPath());
         }
 
         ProdottoDAO.doSave(p);
